refactor: organize nfe types and utils
