+++ src/dfe/nfe/types/consultaCadastro.d.ts
@@ -1,172 +1,4 @@
-import type { LiteralStringUnion } from "@/utils/types";
-
-/**
- * @description Ambiente de emissão da NF-e (tpAmb).
- * producao -> tpAmb = 1
- * homologacao -> tpAmb = 2
- */
-export type Environment = "producao" | "homologacao";
-
-export type UF = LiteralStringUnion<
-  | "AC"
-  | "AL"
-  | "AP"
-  | "AM"
-  | "BA"
-  | "CE"
-  | "DF"
-  | "ES"
-  | "GO"
-  | "MA"
-  | "MT"
-  | "MS"
-  | "MG"
-  | "PA"
-  | "PB"
-  | "PR"
-  | "PE"
-  | "PI"
-  | "RJ"
-  | "RN"
-  | "RS"
-  | "RO"
-  | "RR"
-  | "SC"
-  | "SP"
-  | "SE"
-  | "TO"
->;
-
-/**
- * @description Código numérico de uma UF
- */
-export type UFCode = LiteralStringUnion<
-  | "12"
-  | "27"
-  | "16"
-  | "13"
-  | "29"
-  | "23"
-  | "53"
-  | "32"
-  | "52"
-  | "21"
-  | "51"
-  | "50"
-  | "31"
-  | "15"
-  | "25"
-  | "41"
-  | "26"
-  | "22"
-  | "33"
-  | "24"
-  | "43"
-  | "11"
-  | "14"
-  | "42"
-  | "35"
-  | "28"
-  | "17"
->;
-
-/**
- * @description Servidores de autorização da NF-e
- */
-export type AuthServer =
-  | "AM"
-  | "BA"
-  | "GO"
-  | "MG"
-  | "MS"
-  | "MT"
-  | "PE"
-  | "PR"
-  | "RS"
-  | "SP"
-  | "SVAN"
-  | "SVRS"
-  | "SVCAN"
-  | "SVCRS"
-  | "AN";
-
-export type WebService =
-  | "NfeInutilizacao"
-  | "NfeConsultaProtocolo"
-  | "NfeStatusServico"
-  | "NfeConsultaCadastro"
-  | "RecepcaoEvento"
-  | "NFeAutorizacao"
-  | "NFeRetAutorizacao"
-  | "NFeDistribuicaoDFe";
-
-export type WebServiceUrls = {
-  [env in Environment]: {
-    [server in AuthServer]: {
-      [service in WebService]?: string;
-    };
-  };
-};
-
-export interface GetWebServiceUrlOptions {
-  uf: UF;
-  service: WebService;
-  env: Environment;
-  contingency?: boolean;
-}
-
-export interface NfeRequestOptions<Body> {
-  body: Body;
-  timeout: number;
-}
-
-type WithXmlns<T = unknown> = { "@_xmlns": string } & T;
-type WithVersao<T = unknown> = { "@_versao": string } & T;
-type WithXmlnsVersao<T = unknown> = WithXmlns<WithVersao<T>>;
-
-export type NfeWebServiceResponse<Status, Raw> = {
-  status: Status | "outro";
-  description: string;
-  raw: Raw;
-};
-
-export type NfeStatusServicoRequest = WithXmlns<{
-  consStatServ: WithXmlnsVersao<{
-    tpAmb: "1" | "2";
-    cUF: UFCode;
-    xServ: "STATUS";
-  }>;
-}>;
-
-export type NfeStatusServicoStatus =
-  | "operando"
-  | "paralisado-temporariamente"
-  | "paralisado";
-
-/**
- * @description Resposta completa da consulta de status serviço.
- *
- * @property tpAmb - Tipo de ambiente: 1-Produção 2-Homologação.
- * @property verAplic - Versão do aplicativo que processou a consulta.
- * @property cStat - `107->"operando"`, `108->"paralisado-temporariamente"`, `109->"paralisado"`.
- * @property xMotivo - Descrição da resposta.
- * @property cUF - Código da UF que atendeu a solicitação.
- * @property dhRecbto - Data e hora do processamento.
- * @property [tMed] - Tempo médio de resposta do serviço (em segundos).
- * @property [dhRetorno] - Data e hora previstas para o retorno do serviço.
- * @property [xObs] - Informações adicionais para o contribuinte.
- */
-export interface NfeStatusServicoResponseRaw {
-  tpAmb: "1" | "2";
-  verAplic: string;
-  cStat: string;
-  xMotivo: string;
-  cUF: string;
-  dhRecbto: string;
-  tMed?: string;
-  dhRetorno?: string;
-  xObs?: string;
-}
+import type { WithXmlns, WithXmlnsVersao, UFCode } from "./index";
 
 /**
  * @description Opções para configurar o web service de consulta cadastro.
